Allow tests to be skipped via environment variables
